Update prefixarray.pas
jezis, this is harder than I first thought if not using brute force. I'm not sure how you're supposed to solve this in the time given if you're not aware a) this isn't easy, b) it's been solved using an algorithm, but unless you know that algorithm (i.e. Kidanes) -- good luck.