Update prefixarray.pas
further dribblings....I'm trying not to cheat, find something better than brute force. and my next best methodology is at the bottom commented..... so the idea is first of all you prefix the array. Then you scan the array picking the lowest number to start with. might be -20, for example. Then you count from -20, to the next lowest number lower than -20......do an average on it, then add it to the current average....lol, I'm really grapsing here. brain meltdown..