prefix min slice problem workings
Lots of stuff going on here. I've already nicked the C++ code which solves it in O(n) time and changed it a bit for working in Delphi, but I've also investigated how to do it with prefix sums, which I think will run in O(n) + O(n) time. i.e. one for the prefix, one for the calculations of triplets. Once you have the prefix, you can easily calculate the a 2 element or 3 element set