Update MaxProductThree
codility solution of max product of three in Pascal. 

https://app.codility.com/demo/results/trainingNTDJBF-HJ2/  

The reason I did it like this was to see if I could one pass of the array (without resorting to a traditional sort). I wanted to find out what that would involve, and what the performance in terms of O(n) would be. I'm kinda suprised it runs in O(n * log N)