Update uTextCircle
Added in a typer decorator class, which I'll fill in over the next few days.
Basically, I don't want to go down a rabbit hole of inheritence for different typers.
So, for example, probably the first decorator will just add in a cursor. 
And then I'll start chaining them together.