Update uTextCircle
I've redone this a different way now. One of the problems with the other code...how do you display different parts of text in different color schemes? You can't unless you have lots of seperate typers. With this approach, I request where the text is going to, and allow a controller on the back-end to tinker with the way the text is actually displayed in terms of color..etc