Update uTextCircle
added in TextCursor - basically an edit box which displays a cursor. so that when the typer types letters, there is a cursor displayed. Right, so onto the next thing, parsing....  

So I want now to have 1 sentence split across different types of Text Cursor components - so I can style the letters and words. Easy 
<sigh>
When this is finished it'll be pretty groovy