Create INodeStacks
workable interface for a stack, but an unusual stack, where any object can be inserted into it, as long as it implements this interface. 
It has similar functionality to the spring4D framework stack, in that it can return instances of other stacks of itself....this will make more sense when I get a short demo working.  The reason why I don't use Spring4D? None really, I just didn't want to have to import the entire library.