Create TypeWriter
My version 2 of the other typer...I've had this working perfectly, except for a minor issue with the design time persistence, nothing major, but I've also broken it, as I've added in a script load mechanism, and I have now introduced a text parser (in a different unit).
This has also meant me rethinking this code a bit - but all good... It's like pandora's box. You get to one point, and think groovy, then, what if I had x,y,z and then you realise oh, but that means I have to keep track of the x,y co-ords of the text (given different fonts, and colors) which leads me to a stack kind of set-up....All good - very pleased with results thus far.

A work in progress - but getting there.